requirements.txt
Flask==2.0.1
gspread==5.7.2
oauth2client==4.1.3
python-dotenv==0.19.0
Flask-Login==0.5.0
Werkzeug==2.0.1 